test: separate publish and upgrade tests also rename extension to prepare for native node test runner
