+++ package.json
@@ -15,8 +15,8 @@
     "prebuild": "npm run clean",
     "build": "tsc --build",
     "pretest": "npm run lint && npm run build",
-    "test": "npm run test:spec",
-    "test:spec": "mocha \"./dist/test/**/*.spec.js\"",
+    "test": "npm run test:only",
+    "test:only": "mocha \"./dist/test/**/*.test.js\"",
     "lint": "eslint .",
     "prepack": "npm run build",
     "prettify": "npx prettier . --write"
